initial refactor, seperating the scrim service into two distinct services, add todo doc for further work needed
