+++ src/commands/scrims/signup/sign-up.ts
@@ -1,6 +1,6 @@
 import { MemberCommand } from "../../command";
 import { CustomInteraction } from "../../interaction";
-import { ScrimSignups } from "../../../services/signups";
+import { SignupService } from "../../../services/signups";
 import { isGuildMember } from "../../../utility/utility";
 import { ScrimSignup } from "../../../models/Scrims";
 import { Player } from "../../../models/Player";
@@ -8,7 +8,7 @@ import { PrioService } from "../../../services/prio";
 
 export class SignupCommand extends MemberCommand {
   constructor(
-    private signupService: ScrimSignups,
+    private signupService: SignupService,
     private prioService: PrioService,
   ) {
     super("signup", "Creates a new scrim signup");
@@ -81,7 +81,7 @@ export class SignupCommand extends MemberCommand {
     interaction: CustomInteraction,
     signup: ScrimSignup,
   ) {
-    const scrim = this.signupService.getScrim(interaction.channelId);
+    const scrim = this.signupService.getScrim(interaction.channelId); // Wait, SignupService needs getScrim?
     if (!scrim) {
       console.error(
         "Unable to get applicable prio on a signup because there is no scrim for this channel",
