+++ test/commands/scrims/signup/current-position.test.ts
@@ -7,13 +7,13 @@ import {
 import SpyInstance = jest.SpyInstance;
 import { CustomInteraction } from "../../../../src/commands/interaction";
 import { CurrentPositionCommand } from "../../../../src/commands/scrims/signup/current-position";
-import { ScrimSignupMock } from "../../../mocks/signups.mock";
 import { StaticValueService } from "../../../../src/services/static-values";
 import { StaticValueServiceMock } from "../../../mocks/static-values.mock";
-import { ScrimSignups } from "../../../../src/services/signups";
 import { ScrimSignup } from "../../../../src/models/Scrims";
 import { GetSignupsHelper } from "../../../../src/commands/utility/get-signups";
 import { Player } from "../../../../src/models/Player";
+import { SignupService } from "../../../../src/services/signups";
+import { SignupServiceMock } from "../../../mocks/signups.mock";
 
 describe("Get current position", () => {
   let basicInteraction: CustomInteraction;
@@ -26,7 +26,7 @@ describe("Get current position", () => {
   let getSignupsSpy: SpyInstance<
     Promise<{ mainList: ScrimSignup[]; waitList: ScrimSignup[] } | undefined>,
     [
-      signupsService: ScrimSignups,
+      signupsService: SignupService,
       staticValueService: StaticValueService,
       interaction: CustomInteraction,
     ],
@@ -35,7 +35,7 @@ describe("Get current position", () => {
 
   let command: CurrentPositionCommand;
 
-  const mockSignpuService = new ScrimSignupMock();
+  const mockSignpuService = new SignupServiceMock();
   const mockStaticValueService = new StaticValueServiceMock();
 
   beforeAll(() => {
@@ -59,7 +59,7 @@ describe("Get current position", () => {
     editReplySpy.mockClear();
     getSignupsSpy.mockClear();
     command = new CurrentPositionCommand(
-      mockSignpuService as unknown as ScrimSignups,
+      mockSignpuService as unknown as SignupService,
       mockStaticValueService as StaticValueService,
     );
   });
